Adds generalized window area API's
Introduces API's to get information about window areas
that are available on a device to both move and present
additional content to. The first version of window areas
are rear facing areas on foldable devices. These API's can
be easily expanded to introduce different types of
Window area's to better support use-cases in the future.

Bug: 269660860
Test: WindowAreaControllerImplTest
Relnote: Add RearDisplayPresentationMode APIs

Change-Id: I0401c4abe23a95620e35808000146da11da06faa