Fix Robolectric Compose integration for >1 iteration of trampolining.
aosp/2374993 was the first part of this fix, but only fixed for a single
iteration of compose -> main queue -> compose trampolining. If that
second chunk of compose work posts more work to the main queue, it would
still be broken. So I made the test more rigorous by performing multiple
iterations, and reverted my initial change to RobolectricIdlingStrategy,
and instead moved the fix into ComposeIdlingResource so isIdleNow
returns false not only if there was no more compose work at return time,
but also if it ended up performing any compose work at all.

Bug: b/264548763
Test: RobolectricComposeTest
Change-Id: I34c91ac2cbc5352d4565b4906d80b2a289bce769