Add popBackStack support for exact routes
NavController.popBackStack(route) now supports routes with arguments partially or fully filled in, as long as the filled in arguments matches exactly the entry's filled in arguments.

This method defaults to original behavior of route matching through route pattern. If that fails, it recognizes the provided route as having
arguments filled in and will attempt to match with exact arguments.

Test: ./gradlew navigation:navigation-runtime:cC
Bug: 257514373
Relnote: "NavController.popBackStack(route) now supports routes with arguments partially or fully filled in. Note that the arguments has to be an exact match with the entry's arguments."
Change-Id: I3b37b5d2dd9423bf095cce5632f5208abfe49d70