Adds setLargeIcon(Icon) API to NotifCompat.Builder
Adds setLargeIcon(Icon) API to NotificationCompat.Builder.
Also makes public the previously private function that scales Bitmaps to
the appropriate size for use as LargeIcons; this is required only for
API versions 23 to 27, but cannot be done once a bitmap is wrapped in an
Icon (as the underlying Bitmap is private).

Also added a unit test covering the scaling code in
setLargeIcon(Bitmap) since it was previously untested.

Bug: 172282791
Test: Unit tests
Change-Id: Ic0a5bd849209bd92eba1fbed430369ff30149d08
Relnote: Adds NotifCompat.Builder.setLargeIcon(Icon) signature