Deprecate Rect.toAndroidRect
Rect.toAndroidRect is a lossy conversion (from a floating point type to
an integral type), but the method name does not indicate that a lossy
conversion is taking place. This is compounded by the fact that the
Android Rect and the Compose Rect share a name, but not the underlying
representation (the Compose integral rectangle type is IntRect, whereas
the floating point Android rectangle type is RectF).

Rect.toAndroidRect is now deprecated, with the inlined implementation as
the replacement.

All current usages in Compose are migrated away from Rect.toAndroidRect,
with the inlined implementation, ensuring no behavior changes in this
CL.

Fixes: 265811631
Tests: Existing
Change-Id: I5f15668677afdaba60a6123ea4899c2e62d0fba9