Improve test for ProcessGlobalConfigActivity
ProcessGlobalConfigActivity runs in a process different from that of the
MainActivity. Since basic instrumentation test has synchronization
guarantees limited to a single process, it was impossible to wait for
the WebView to be rendered in the new process. To work around this,
Thread.sleep() was used which lead to flaky test runs.

This CL adds Multiprocess Espresso support to the webkit instrumentation tests,
which allows waiting for a View to be rendered on a different process
belonging to the same app. This is currently made use of in the
ProcessGlobalConfigActivity test and is now available for other
instrumentation tests within androidx.webkit.

Multiprocess Espresso documentation:
https://developer.android.com/training/testing/espresso/multiprocess

Bug: 263348639
Test: ./gradlew webkit:integration-tests:testapp:connectedAndroidTest
Change-Id: I4aeef9320ac76ce70245e8e437b73ccadd4667e9