Add a RequestDisplayInfo field to the CreateCustomCredentialRequest.
During the UI stage, we need the basic user credential information for
display purposes. The info has to come directly from the request and the
providers will not get them because the user hasn't consented yet when
we surface the UI. We need all custom request impls to provider this
information so that we can display it properly.

Ideally we should require this value and mark it as non-null. However, we can't do so because it's always possible (though extremely unlikely and discouraged) that someone construct a custom request directly from the framework api where this structured data structure is absent. We still want to allow that case to flow through, hence leaving the field as nullable so that parsing doesn't crash when this property isn't detected.

Bug: 266152799
Test: credentials/credentials/src/androidTest/java/androidx/credentials/*Test
Relnote: "Add a RequestDisplayInfo field to the CreateCustomCredentialRequest"
Change-Id: Ia6e9b6f3b9bf76612d6c7706db30b3a21e5727ca