Expandable components and demos
Add two new material demos to showcase how to use an Expandable
control, and the components to support them.
One to show/hide several items in a Scaling Lazy Column, and one to
show more content in a Text.

Relnote: "Added implementations of a two expandable components (one for a
group of items in a ScalingLazyColumn, and one for a single item.)"
Test: N/A
Bug: 265931525

Change-Id: I95dd5f8d5e1d76ee6cb08b176b109d74c0c543ef