Add prepareForTransition API to NavigatorState
There is currently no way for NavigatorState to address the intermediate
states during transitions. We should add an API that allows
NavBackStackEntries to not need to skip over states, specifically when
entries are being popped off the back stack.

RelNote: "`NavigatorState` now provides the `prepareForTransition` API to
allow the `Navigator` to move NavBackStackEntries to intermediate
`Lifecycle.State`s."
Test: TestNavigatorStateTest
Bug: 238686802

Change-Id: I42c214311d1ab67f589f5749a62133ec564044f4