Added androidx extension flag to disable all tests in generated xml configuration
In aosp/2305556 we need to test the baseline profile gradle plugin.
CI is failing because the target apk is not installed. This is a use
case similar to macrobenchmark for which we've some ad-hoc configuration.
In this case though, the final goal is to test the gradle plugin so it's
not worth to run only a piece of the architecture. This PR introduces a
param in the `androidx` extension to disable all the tests in the
generated xml configuration.

Test: ./gradlew :buildSrc-tests:test
Bug: 266630020
Change-Id: Iab7bc0848c54aa0e0c23b5d626749d8e3ce48b01