Allow nullable locking directory in SupportSQLiteLock.
Some family and mostly older devices don't properly create the cache directory and a null value can be returned by the context, to protect from such case the lock directory for process locking a database is not obligatory and the behaviour will fallback as it used to before the process lock, i.e. only in-memory locking.

Bug: 265353349
Test: Existing
Change-Id: I13f5e6253afa33bdea69559b40f43298123cbb0d