Improve test flakiness for layout tests.
The AppWidgetHostView adds its own default padding. The previous version
of this test tried to 0 out the padding. This doesn't really work because
the view reapplies the padding in various different places internally.

The fix here is to make the host bigger by the padding amount, and allow
the inner view (the widget we care about) to be the provided size naturally

Bug: 266588723
Bug: 265078768
Test: Ran AppWidgetReceiverTests
Change-Id: I1655f166a17bc2242396bbbf1b673107ff641e21