In SurfaceProcessorNode, make sure identical OutConfig returns different edges.
In the processing pipeline, we may configure the node to produce 2 identical outputs with the same crop, size and rotation. The issue is, their OutConfig have the same hashCode() which overwrites one another in HashMap, so that only one will return.

In this CL, we add a UUID field to OutConfig and force the same config have different hash code.

Bug:265818567
Test: manual test and ./gradlew bOS
Change-Id: I7d424a5cf2c8efc5b55124597c88610ee9692715