Increment DataStore version "after" write
Important changes:
* writes increment the counter after the write. This ensures that the reader
can still use the version it has for up-to-date checks.
Lets say data on disk is A, version 0 (A, V0) and it is being updated to B.

writer will start the write at V0, commit the change and then set to V1

Reader will read the version first then read the data and then re-read the
version if it has the lock. In this case, reader can only see:
version 0 data A
version 0 data B
version 1 data B

notice that in each case, version is equal or less than the version on disk
which is good enough to order cached values.

Without this change (when incremen happens before the write), reader might see
version 1, data A (new version, old data); which will break ordering.

* In memory cache is moved to a helper class to control updates (ensure they
are syncronized and also never go back in version)

* Initialization is moved to a separate class which ensures it is runs at most once
to completion (this is just a cleanup to isolate things)

Bug: 267815238
Bug: 267836698
Fixes: 242765757
Test: MultiProcessDataStore.stressTest

Change-Id: I6ae5a8c067ac8ee949a7481e50bbe82ae0054084