Set initial value on distinctUntilChanged()
If the LiveData you're calling distinctUntilChanged()
already has a value set (checked via
isInitialized), we should immediately set the value
on the output LiveData, thus avoiding any time
where the value is not set.

This fixes cases such as with Compose's observeAsState
which look at whether the LiveData has been
initialized in order to know what State to expose
in the first composition as of aosp/2433877

Relnote: "The `LiveData.distinctUntilChanged()` extension
now sets the `value` of the returned `LiveData` if the
previous `LiveData` has had a value set on it. This does
not change the observation behavior - updated values from
the source LiveData will still only apply once you start
observing the LiveData returned from `distinctUntilChanged()`."
Test: new initialValueIsSet test
BUG: 214434197

Change-Id: Ib482f0cc74fad451e0ff37088a11fcf2682f0f8a