Submit repeating requests in order
Currently when GraphProcessor is given a repeating request, it launches
a coroutine with the lightweight dispatcher in order to submit the
last repeating request. That means there can be multiple parallel
coroutines attempting to submit repeating requests, and it is possible
then for them to be submitted out-of-order.

This CL fixes it by doing the following:
1. Establish a repeating request queue to ingest incoming requests.
2. Make sure only 1 coroutine tries to submit requests at any time.
3. Drop stale requests that preceded a succeeded request.

Bug: 267709383
Test: CameraPipe unit tests
Change-Id: I59835c47166ab24b669793270c8006be1d232c69