Benchmark tests for PickerGroup
Refactored the test cases for ScalingLazyColumn since PickerGroup is also based out of SLC. We should be able to reuse the methods in Picker as well or other components based out of SLC.

On a different note, we may be able to get rid of the extension methods since the original bug which caused the issue has been fixed. Since that was not the scope of this bug (making changes in SLC), I have instead refactored the test cases right now. Created a bug for tracking the changes in SLC. b/270318565

Test: androidx.wear.compose.material.benchmark.PickerGroupBenchmark
Bug: 264147685
Change-Id: I90434c56e73b0f02455853a08c5eb44448feec59