Change CVF fix for parameter type implicit casting
aosp/2430534 made it so if the real types of the arguments to a CVF fix method didn't match the expected types, the fix included an explicit cast. However, this was a redundant cast which won't actually fix the problem (https://android-review.git.corp.google.com/c/platform/frameworks/support/+/2443359/comment/647abf4b_62fb9311/).

This changes the approach -- if the type of an argument and parameter don't match and would be an invalid implicit cast, the generated fix method now uses the argument type, instead of adding a cast at the call site to the parameter type.

This doesn't change the type of a varargs parameter, which means if there's an invalid implicit cast with varargs, the implicit cast detector will add another autofix on top of this one, which isn't ideal.

Test: Updated tests from aosp/2430534 with the new expected behavior, and added a varargs test for the implicit cast fix
Bug: 266845827
Change-Id: Icc14f513be8059a23204d25e4f5f453ca3181a77