Improves LocalConfiguration to compare instances and reduce invalidation
Previously we always invalidated LocalConfiguration whenever we recompose at the provider: this was made because the configuration object will be mutated internally, and so we cannot compare it against itself to see if it has changed. However, instead we can just maintain a separate deep copy and compare that to the mutated instance when it changes, so we only invalidate the hierarchy if the new configuration is actually different from before.

Fixes: b/271286350
Test: AndroidCompositionLocalTest
Change-Id: I7a9240329352ba285971f6946c29b673316644ae