Set minimum SQL statement cache during migrations.
This is a workaround of b/183028015 where the framework's connection is caching prepared statements using the SQL query string as key and will cause issues if the schemas changes since the cache is not cleared and the statement might contain outdated information. By setting the cache size to the minimum allowed (1) we can reduce (but not completely eliminate) the chance of a SELECT query being used from cached whose backing schemas has changed. Note that the reduction in cache size is only temporary and during migrations where the schema is most likely to change.

Bug: 271083856
Test: FrameworkSQLiteDatabaseTest
Change-Id: I35ed1cecf7837d4d218a6af842edd1deac53e1e2