Recover from ERROR_CAMERA_DEVICE
ERROR_CAMERA_DEVICE is considered recoverable in CameraX, and indeed it
is recoverable if we receive onCameraAvailable() from Camera2.

This CL changes the restart mechanism in CameraControllers such that
GraphLifecycleManager also feeds the reason for the restart to
CameraControllers. CameraControllers can then utilize this to determine
whether they should restart.

We also correct a minor mistake in which camera open exceptions were
incorrectly translated to ERROR_OTHER_RECOVERABLE_ERROR, when in fact
they should be translated to ERROR_CAMERA_FATAL_ERROR.

Bug: 263201241
Test: CameraPipe unit tests
Change-Id: Ib2ca5d8f513234cd20b7bd51f26ffa82e29c742d