Fix binding to parameterless watchface
A parameterless watch face is one that has booted without any
start up params. Typically this can only happen if a WSL watchface
is upgraded to an androidx one, so WallpaperManager knows about it
but WearServices/WSL does not.

This patch allows a subsequent call to
getOrCreateInteractiveWatchFaceClient to bind to the engine.

Bug: 268472655
Test: New unit test
Change-Id: Ic037e1e3f9659f1368cc843727059fd9d3fcf37b