Extract logic that decides whether or not a SemanticsNode (and its AccessibilityNodeInfo) will be `screenReaderFocusable`.
Test: Existing tests (no-op refactoring)

When `currentSemanticsNodes` are retrieved, their order is sorted and saved in an `idToBefore` map. Then when ANIs are created, each one is individually populated: `screenReaderFocusable` is determined based on other properties, and then the `traversalBefore` and `traversalAfter` properties are also set. But we eventually want to sort nodes only if they are either screenReaderFocusable or a traversalGroup. So the logic that determines whether an ANI is `screenReaderFocusable` or not must be taken out of the `populateAccessibilityNodeInfoProperties`, which is what this CL does.

Change-Id: If55391068219988282cd01d9028b9da7f79cb4bf