Prepare all entries in change started listener during pop
When the fragment manager does a pop, we should not only prepare the
RESUMED entry, but also any entries that might be preparing to move as
well. This ensures that the entry coming back from the back stack will
be STARTED instead of CREATED.

This should currently be a no-op, but will be needed for future
features.

Test: all tests pass
Change-Id: I356abb41499f5cc9b52e22e7b7b6a96977284297