Migrate IME visible fix for AppCompatTextView
Fixes an issue where if the mServedView gets detached,
but a different view was already set as the mNextServedView,
the next served view gets mistakenly cleared as well.

In CL[1] for AppCompatEditText, update the served view in time
when the view is detaching. Also handle this in AppCompatTextView
if it's an editable TextView.

[1]: I77ba52264764fc0716f0ec0be901ecf8c74e01b2

Bug: 258048232
Test: AppCompatImeFocusTest
Change-Id: I562915aa8bb376be22370be6b47e0486716f1db6