ComplicationDataExpressionEvaluator closes on Flow termination.
Removes the necessity of a close() method by scoping the evaluation
inside the Flow's lifecycle.

This makes for a cleaner API, where dependencies are provided to the
constructor, but the data to evaluate is provided to evaluate(). It also
prevents the potential mistake of calling evaluate() or collecting the
Flow multiple times.

Cleanup test in aosp/2496942 because it requires a DynamicTypeEvaluator
API that shows resource cleaning, e.g. TimeGateway.

Bug: 273978322
Change-Id: If184bb4e5aa6e6905132ecec21dc1b2508be9211