Revert "Revert "[CameraPipe] Add lock behaviors for all 3A modes for startFocusAndMetering""
This reverts commit 859d89c7d16280b2b7beed925098836083242ca1.

Reason for revert:
It seems that re-applying the reverted CL on LKGB finds no flakiness (result in b/274588483). So, the bug is probably not related with the reverted CL and was caused/fixed by something else.

Also, the whole AudioSourceTest has been ignored temporarily in aosp/2504915 and the root cause is being investigated in b/274840083

Change-Id: If4747d00aeec6c5b4627337429dea84f3cc5820a