Always use a temp local val in FieldGetter when gen Kotlin and nullable.
Otherwise compilation will fail due to an impossible smart cast when the entity property is from another module (already compiled), even if the property is immutable (val).

Even though the temp local val is not necessary when the call type is a field and immutable, this change still generates it since otherwise we would need to check or expose through XProcessing the KSP declaration's origin.

Bug: 274760383
Test: DaoKotlinCodeGenTest
Change-Id: Ia48f834afa8569b3cb8c790dcba47455ce04c990