Fix Navigation test to ensure they are no longer flaky
Using Animators in our tests was causing them to be flaky since we were
checking for specific states during the animation and on the remote
tests devices have Animators turned off.

Instead of relying on being able to test only the state when the
Animators are running, we can just check both to ensure that we can pass
tests in both cases.

Fixes: 271634544
Fixes: 269297210
Fixes: 271634544
Test: this edits tests
Change-Id: I185af180d5a4022b8dee01daaf319fbdad900cbf