Prevent accidentally removing active fragments
It is possible with saving and restoring state to accidentally make
fragment manager mark an active fragment as inactive.

This results in a failed restore after a config change.

We should only make FragmentStateManagers that are active inactive. This
ensures that active and inactive calls are 1 to 1.

RelNote: "Fixed a bug in FragmentManager where saving and restoring in
the same frame could cause a crash."
Test: Added FragmentManagerTest
Bug: 246519668

Change-Id: Ib36aff4a7263e10d7d795b4d8dbd6431d2552e5c