Introduce BasicSecureTextField in text2
Add a new Experimental text input composable to foundation that's specifically designed to be used as a password field.

`BasicSecureTextField` accepts 3 methods of obfuscation of its content; `Visible`, `RevealLastTyped`, and `Hidden`.

This CL also adds another Experimental API to `BasicTextField2` called `CodepointTransformation`. This new interface is targeted to be a replacement of `VisualTransformation` in new text fields.

`OffsetMapping` has been a huge source of pain for both developers and maintainers. Allowing text length to change arbitrarily has been proven to be difficult for many reasons. Writing a correct function that calculated OffsetMapping has been a very difficult task. Also, selection and text offsets do not directly translate to each other which makes selection act wonky even when `TransformedText` is calculated correctly.

`CodepointTransformation` requires a 1-to-1 mapping of each codepoint in original text, which guarantees that the length of the current input and rendered text are always the same. This makes sure that selection does not run into any weird offset misalignment.

Test: :compose:foundation:foundation:cAT
Relnote: N/A
Change-Id: I4fbec661baa9cf8ba22e798f221acb24757e1e3e